Integrating the Universal Downloader with the Imgur API
Also can't test this one because I don't have WiFi to download the imgur
python library or the memes themselves. Anyone willing to test it, be my
guest and change what you need.
A note: the imgur 'text0' field (which in the Reddit API I used as the
filename; perhaps we need a field for that in our JSON standard) needs
to be changed. I don't know how the Imgur Python library works because I
don't have WiFi to check.
Also, if anyone is willing to send me a note on how Haskell can handle
JSON parsing, please do so. I will look into running Haskell from Python
(because the python file in that folder hasn't been touched)