Changed datePulled data type and Updated the JSON to fit the new standard
The new standard being 'imageUrl' instead of 'url', of course.
Also, I change the type of datePulled from int to float so we can be
more accurate. However, it still converts it to an int when it uses the
date as a file name, but that's only so that it doesn't have too many
periods in the filename.