Fixed some things
The MemeGenerator Haskell code was appending to the JSON file and not
overwriting it. Fixed that by deleting everything in that file on start.
Added a filename field to the JSON standard. This will help to
distinguish things as I was using the text0 field for the filename.
Will implement different folders for the MemeGenerator code (because the
title will tell us what kind of meme it is, so I will separate the memes
by that field).
Implemented a cross platform command (as long as you have Haskell and
Haskell Stack installed) for running the MemeGenerator Haskell code
(using runhaskell instead of a compiled binary)