Changed code
Changed the file names for all of the libraries to main and changed the
downloader code to match that.
Added the functionality to search for the subreddit file in the root as
well as in the reddit folder. If it doesn't find it in either, then it
raises an exception (which might not work, not sure).